TradeLogIB.py 4.0: added support for modern TWS versions which report time in UTC format. TradeLogIB now automatically converts trade time to Local Time Zone; TradeLogIB.py: added new --UTC command line option which allows to save trades in UTC format; TradeLogIB.py: added two new dependencies: pytz, tzlocal; ib_insync: latest ib_insync code which is guaranteed to work correctly with the script.

+++ ib_insync/ibcontroller.py
@@ -4,10 +4,11 @@ import logging
 import configparser
 from contextlib import suppress
 
+from eventkit import Event
+
 from ib_insync.objects import Object
 from ib_insync.contract import Forex
 from ib_insync.ib import IB
-from ib_insync.event import Event
 import ib_insync.util as util
 
 __all__ = ['IBC', 'IBController', 'Watchdog']
@@ -17,73 +18,63 @@ class IBC(Object):
     """
     Programmatic control over starting and stopping TWS/Gateway
     using IBC (https://github.com/IbcAlpha/IBC).
-    
-    Arguments:
-    
-    * ``twsVersion`` (required): The major version number for TWS or gateway.
-    * ``tradingMode``: 'live' or 'paper'.
-    * ``userid``: IB account username. It is recommended to set the real
-      username/password in a secured IBC config file.
-    * ``password``: IB account password.
-    * ``twsPath``: Path to the TWS installation folder.
-    
-      =======  ==============
-      Default
-      =======================
-      Linux    ~/Jts
-      OS X     ~/Applications
-      Windows  C:\\\\Jts
-      =======  ==============
-      
-    * ``twsSettingsPath``: Path to the TWS settings folder.
-    
-      ========  =============
-      Default
-      =======================
-      Linux     ~/Jts
-      OS X      ~/Jts
-      Windows   Not available
-      ========  =============
-      
-    * ``ibcPath``: Path to the IBC installation folder.
-    
-      ========  =============
-      Default
-      =======================
-      Linux     /opt/ibc
-      OS X      /opt/ibc
-      Windows   C:\\\\IBC
-      ========  =============
-      
-    * ``ibcIni``: Path to the IBC configuration file.
-    
-      ========  =============
-      Default
-      =======================
-      Linux     ~/ibc/config.ini
-      OS X      ~/ibc/config.ini
-      Windows   %%HOMEPATH%%\\\\Documents\\\\IBC\\\\config.ini
-      ========  =============
-      
-    * ``javaPath``: Path to Java executable.
-      Default is to use the Java VM included with TWS/gateway.
-    * ``fixuserid``: FIX account user id (gateway only).
-    * ``fixpassword``: FIX account password (gateway only).
-    
+
+    Args:
+
+        twsVersion (int): (required) The major version number for
+            TWS or gateway.
+        gateway (bool):
+            * True = gateway
+            * False = TWS
+        tradingMode (str): 'live' or 'paper'.
+        userid (str): IB account username. It is recommended to set the real
+            username/password in a secured IBC config file.
+        password (str): IB account password.
+        twsPath (str): Path to the TWS installation folder.
+            Defaults:
+
+            * Linux:    ~/Jts
+            * OS X:     ~/Applications
+            * Windows:  C:\\\\Jts
+        twsSettingsPath (str): Path to the TWS settings folder.
+            Defaults:
+
+            * Linux:     ~/Jts
+            * OS X:      ~/Jts
+            * Windows:   Not available
+        ibcPath (str): Path to the IBC installation folder.
+            Defaults:
+
+            * Linux:     /opt/ibc
+            * OS X:      /opt/ibc
+            * Windows:   C:\\\\IBC
+        ibcIni (str): Path to the IBC configuration file.
+            Defaults:
+
+            * Linux:     ~/ibc/config.ini
+            * OS X:      ~/ibc/config.ini
+            * Windows:   %%HOMEPATH%%\\\\Documents\\\\IBC\\\\config.ini
+        javaPath (str): Path to Java executable.
+            Default is to use the Java VM included with TWS/gateway.
+        fixuserid (str): FIX account user id (gateway only).
+        fixpassword (str): FIX account password (gateway only).
+
+    This is not intended to be run in a notebook.
+
     To use IBC on Windows, the proactor (or quamash) event loop
     must have been set:
-    
+
     .. code-block:: python
-        
+
         import asyncio
         asyncio.set_event_loop(asyncio.ProactorEventLoop())
-    
+
     Example usage:
-    
+
     .. code-block:: python
-    
-        ibc = IBC(969, gateway=True, tradingMode='live',
-                userid='edemo', password='demouser')
+
+        ibc = IBC(974, gateway=True, tradingMode='live',
+            userid='edemo', password='demouser')
         ibc.start()
         IB.run()
     """
@@ -112,7 +103,7 @@ class IBC(Object):
         Object.__init__(self, *args, **kwargs)
         if not self.ibcPath:
             self.ibcPath = '/opt/ibc' if os.sys.platform != 'win32' \
-                    else 'C:\\IBC'
+                else 'C:\\IBC'
         self._proc = None
         self._monitor = None
         self._logger = logging.getLogger('ib_insync.IBC')
@@ -128,13 +119,13 @@ class IBC(Object):
         """
         Launch TWS/IBG.
         """
-        util.syncAwait(self.startAsync())
+        util.run(self.startAsync())
 
     def terminate(self):
         """
         Terminate TWS/IBG.
         """
-        util.syncAwait(self.terminateAsync())
+        util.run(self.terminateAsync())
 
     async def startAsync(self):
         if self._proc:
@@ -143,7 +134,8 @@ class IBC(Object):
 
         # create shell command
         win32 = os.sys.platform == 'win32'
-        cmd = [f'{self.ibcPath}\\scripts\\StartIBC.bat' if win32 else
+        cmd = [
+            f'{self.ibcPath}\\scripts\\StartIBC.bat' if win32 else
             f'{self.ibcPath}/scripts/ibcstart.sh']
         for k, v in self.dict().items():
             arg = IBC._Args[k][2 if win32 else 1]
@@ -156,19 +148,25 @@ class IBC(Object):
                     cmd.append(str(v))
 
         # run shell command
-        self._proc = await asyncio.create_subprocess_exec(*cmd,
-                stdout=asyncio.subprocess.PIPE)
+        self._proc = await asyncio.create_subprocess_exec(
+            *cmd, stdout=asyncio.subprocess.PIPE)
         self._monitor = asyncio.ensure_future(self.monitorAsync())
 
     async def terminateAsync(self):
         if not self._proc:
             return
         self._logger.info('Terminating')
-        self._monitor.cancel()
-        self._monitor = None
-        with suppress(ProcessLookupError):
-            self._proc.terminate()
-            await self._proc.wait()
+        if self._monitor:
+            self._monitor.cancel()
+            self._monitor = None
+        if os.sys.platform == 'win32':
+            import subprocess
+            subprocess.call(
+                ['taskkill', '/F', '/T', '/PID', str(self._proc.pid)])
+        else:
+            with suppress(ProcessLookupError):
+                self._proc.terminate()
+                await self._proc.wait()
         self._proc = None
 
     async def monitorAsync(self):
@@ -182,17 +180,18 @@ class IBC(Object):
 class IBController(Object):
     """
     For new installations it is recommended to use IBC instead.
-    
+
     Programmatic control over starting and stopping TWS/Gateway
     using IBController (https://github.com/ib-controller/ib-controller).
-    
+
     On Windows the the proactor (or quamash) event loop must have been set:
-    
+
     .. code-block:: python
-        
+
         import asyncio
         asyncio.set_event_loop(asyncio.ProactorEventLoop())
 
+    This is not intended to be run in a notebook.
     """
     defaults = dict(
         APP='TWS',  # 'TWS' or 'GATEWAY'
@@ -225,19 +224,19 @@ class IBController(Object):
         """
         Launch TWS/IBG.
         """
-        util.syncAwait(self.startAsync())
+        util.run(self.startAsync())
 
     def stop(self):
         """
         Cleanly shutdown TWS/IBG.
         """
-        util.syncAwait(self.stopAsync())
+        util.run(self.stopAsync())
 
     def terminate(self):
         """
         Terminate TWS/IBG.
         """
-        util.syncAwait(self.terminateAsync())
+        util.run(self.terminateAsync())
 
     async def startAsync(self):
         if self._proc:
@@ -257,8 +256,8 @@ class IBController(Object):
         ext = 'bat' if os.sys.platform == 'win32' else 'sh'
         cmd = f'{d["IBC_PATH"]}/Scripts/DisplayBannerAndLaunch.{ext}'
         env = {**os.environ, **d}
-        self._proc = await asyncio.create_subprocess_exec(cmd, env=env,
-                stdout=asyncio.subprocess.PIPE)
+        self._proc = await asyncio.create_subprocess_exec(
+            cmd, env=env, stdout=asyncio.subprocess.PIPE)
         self._monitor = asyncio.ensure_future(self.monitorAsync())
 
     async def stopAsync(self):
@@ -302,54 +301,61 @@ class IBController(Object):
 
 class Watchdog(Object):
     """
-    Start, connect and watch over the TWS or gateway app to keep it running
-    in the face of crashes, freezes and network outages.
-    
+    Start, connect and watch over the TWS or gateway app and try to keep it
+    up and running. It is intended to be used in an event-driven
+    application that properly initializes itself upon (re-)connect.
+
+    It is not intended to be used in a notebook or in imperative-style code.
+    Do not expect Watchdog to magically shield you from reality. Do not use
+    Watchdog unless you understand what it does and doesn't do.
+
+    Args:
+        controller (Union[IBC, IBController]): (required) IBC or IBController
+            instance.
+        ib (IB): (required) IB instance to be used. Do no connect this
+            instance as Watchdog takes care of that.
+        host (str): Used for connecting IB instance.
+        port (int):  Used for connecting IB instance.
+        clientId (int):  Used for connecting IB instance.
+        connectTimeout (float):  Used for connecting IB instance.
+        appStartupTime (float): Time (in seconds) that the app is given
+            to start up. Make sure that it is given ample time.
+        appTimeout (float): Timeout (in seconds) for network traffic idle time.
+        retryDelay (float): Time (in seconds) to restart app after a
+            previous failure.
+
     The idea is to wait until there is no traffic coming from the app for
     a certain amount of time (the ``appTimeout`` parameter). This triggers
     a historical request to be placed just to see if the app is still alive
     and well. If yes, then continue, if no then restart the whole app
     and reconnect. Restarting will also occur directly on error 1100.
-    
-    Arguments:
-    
-    * ``controller`` (required): IBC or IBController instance;
-    * ``ib`` (required): IB instance to be used. Do no connect this
-      instance as Watchdog takes care of that.
-    * ``host``, ``port``, ``clientId`` and ``connectTimeout``: Used for
-      connecting to the app;
-    * ``appStartupTime``: Time (in seconds) that the app is given to start up;
-      Make sure that it is given ample time;
-    * ``appTimeout``: Timeout (in seconds) for network traffic idle time;
-    * ``retryDelay``: Time (in seconds) to restart app after a previous failure;
-      If left empty then Watchdog creates the IB instance.
-    
-    Note: ``util.patchAsyncio()`` must have been called before.
-    
-    Example usage:
-    
-    .. code-block:: python
-    
-        util.patchAsyncio()
 
-        ibc = IBC(969, gateway=True, tradingMode='paper')
+    Example usage:
+
+    .. code-block:: python
+
+        def onConnected():
+            print(ib.accountValues())
+
+        ibc = IBC(974, gateway=True, tradingMode='paper')
         ib = IB()
-        app = Watchdog(ibc, ib, port=4002)
-        app.start()
-        print(app.ib.accountValues())
+        ib.connectedEvent += onConnected
+        watchdog = Watchdog(ibc, ib, port=4002)
+        watchdog.start()
         IB.run()
-    
+
     Events:
-        * ``startingEvent(watchdog)``
-        * ``startedEvent(watchdog)``
-        * ``stoppingEvent(watchdog)``
-        * ``stoppedEvent(watchdog)``
-        * ``softTimeoutEvent(watchdog)``
-        * ``hardTimeoutEvent(watchdog)``
+        * ``startingEvent`` (watchdog: :class:`.Watchdog`)
+        * ``startedEvent`` (watchdog: :class:`.Watchdog`)
+        * ``stoppingEvent`` (watchdog: :class:`.Watchdog`)
+        * ``stoppedEvent`` (watchdog: :class:`.Watchdog`)
+        * ``softTimeoutEvent`` (watchdog: :class:`.Watchdog`)
+        * ``hardTimeoutEvent`` (watchdog: :class:`.Watchdog`)
     """
 
-    events = ['startingEvent', 'startedEvent', 'stoppingEvent', 'stoppedEvent',
-            'softTimeoutEvent', 'hardTimeoutEvent']
+    events = [
+        'startingEvent', 'startedEvent', 'stoppingEvent', 'stoppedEvent',
+        'softTimeoutEvent', 'hardTimeoutEvent']
 
     defaults = dict(
         controller=None,
@@ -361,7 +367,7 @@ class Watchdog(Object):
         appStartupTime=30,
         appTimeout=20,
         retryDelay=2)
-    __slots__ = list(defaults.keys()) + events + ['_watcher', '_logger']
+    __slots__ = list(defaults.keys()) + events + ['_runner', '_logger']
 
     def __init__(self, *args, **kwargs):
         Object.__init__(self, *args, **kwargs)
@@ -374,72 +380,83 @@ class Watchdog(Object):
             raise ValueError('IB instance must not be connected')
         assert 0 < self.appTimeout < 60
         assert self.retryDelay > 0
-        self.ib.errorEvent += self.onError
-        self.ib.disconnectedEvent += self.flush
-        self._watcher = asyncio.ensure_future(self.watchAsync())
+        self._runner = None
         self._logger = logging.getLogger('ib_insync.Watchdog')
 
     def start(self):
-        self.startingEvent.emit(self)
         self._logger.info('Starting')
-        self.controller.start()
-        IB.sleep(self.appStartupTime)
-        try:
-            self.ib.connect(self.host, self.port, self.clientId,
-                    self.connectTimeout)
-            self.ib.setTimeout(self.appTimeout)
-            self.startedEvent.emit(self)
-        except:
-            self.controller.terminate()
-            self.scheduleRestart()
+        self.startingEvent.emit(self)
+        self._runner = asyncio.ensure_future(self.runAsync())
 
     def stop(self):
-        self.stoppingEvent.emit(self)
         self._logger.info('Stopping')
+        self.stoppingEvent.emit(self)
         self.ib.disconnect()
-        self.controller.terminate()
-        self.stoppedEvent.emit(self)
+        self._runner = None
 
-    def scheduleRestart(self):
-        loop = asyncio.get_event_loop()
-        loop.call_later(self.retryDelay, self.start)
-        self._logger.info(f'Schedule restart in {self.retryDelay}s')
+    async def runAsync(self):
 
-    def flush(self):
-        self.stop()
-        self.scheduleRestart()
+        def onTimeout(idlePeriod):
+            if not waiter.done():
+                waiter.set_result(None)
 
-    def onError(self, reqId, errorCode, errorString, contract):
-        if errorCode == 1100:
-            self._logger.error(f'Error 1100: {errorString}')
-            self.ib.disconnect()
+        def onError(reqId, errorCode, errorString, contract):
+            if errorCode == 1100 and not waiter.done():
+                waiter.set_exception(Warning('Error 1100'))
 
-    async def watchAsync(self):
-        while True:
-            await self.ib.wrapper.timeoutEv.wait()
-            # soft timeout, probe the app with a historical request
-            self._logger.debug('Soft timeout')
-            self.softTimeoutEvent.emit(self)
-            contract = Forex('EURUSD')
-            probe = self.ib.reqHistoricalDataAsync(
-                    contract, '', '30 S', '5 secs', 'MIDPOINT', False)
+        def onDisconnected():
+            if not waiter.done():
+                waiter.set_exception(Warning('Disconnected'))
+
+        while self._runner:
             try:
-                bars = await asyncio.wait_for(probe, 4)
-                if not bars:
-                    raise Exception()
+                await self.controller.startAsync()
+                await asyncio.sleep(self.appStartupTime)
+                await self.ib.connectAsync(
+                    self.host, self.port, self.clientId, self.connectTimeout)
+                self.startedEvent.emit(self)
                 self.ib.setTimeout(self.appTimeout)
-            except:
-                # hard timeout, flush everything and start anew
-                self._logger.error('Hard timeout')
-                self.hardTimeoutEvent.emit(self)
-                self.ib.disconnect()
+                self.ib.timeoutEvent += onTimeout
+                self.ib.errorEvent += onError
+                self.ib.disconnectedEvent += onDisconnected
+
+                while self._runner:
+                    waiter = asyncio.Future()
+                    await waiter
+                    # soft timeout, probe the app with a historical request
+                    self._logger.debug('Soft timeout')
+                    self.softTimeoutEvent.emit(self)
+                    probe = self.ib.reqHistoricalDataAsync(
+                        Forex('EURUSD'), '', '30 S', '5 secs',
+                        'MIDPOINT', False)
+                    bars = None
+                    with suppress(asyncio.TimeoutError):
+                        bars = await asyncio.wait_for(probe, 4)
+                    if not bars:
+                        self.hardTimeoutEvent.emit(self)
+                        raise Warning('Hard timeout')
+                    self.ib.setTimeout(self.appTimeout)
+
+            except ConnectionRefusedError:
+                pass
+            except Warning as w:
+                self._logger.warn(w)
+            except Exception as e:
+                self._logger.exception(e)
+            finally:
+                self.ib.timeoutEvent -= onTimeout
+                self.ib.errorEvent -= onError
+                self.ib.disconnectedEvent -= onDisconnected
+                await self.controller.terminateAsync()
+                self.stoppedEvent.emit(self)
+                if self._runner:
+                    await asyncio.sleep(self.retryDelay)
 
 
 if __name__ == '__main__':
     asyncio.get_event_loop().set_debug(True)
     util.logToConsole(logging.DEBUG)
-    util.patchAsyncio()
-    ibc = IBC(971, gateway=True, tradingMode='paper')
+    ibc = IBC(974, gateway=True, tradingMode='paper')
 #             userid='edemo', password='demouser')
     ib = IB()
     app = Watchdog(ibc, ib, port=4002, appStartupTime=15, appTimeout=10)
